Upgrades people! Upgrades!
So, stuff happened. Made the documentatin file look nicer, and changed all the new section comments to todo things so they're in the todo tab at the bottom. Also started adding in all those bits into the docs that were missed out. Also, finally and for real this time, fixed #13!!